Add EncryptWithExistingKey
